hw1: Trip related tests + more Route tests

+++ hw1/bus-service-backend/src/main/java/com/tqs108636/busservicebackend/model/Trip.java
@@ -35,7 +35,7 @@ public class Trip {
     private Date departureTime;
 
     @NotNull
-    private int priceEuro;
+    private float priceEuro;
 
     @NotNull
     private int numberOfSeats;
